Rename test_scheduler.py to test_host_appointer_py, and move tests out of host_appointer.py
